attach wallet to message so we can lookup user info.
at first was trying to switch from ed25519 to secp256k1
but it was too annoying... we only have user.wallet
so we don't know the pubkey with a signed txn.

so user could use arbitrary private key, and set any wallet in the message payload.
so the signing is not "real" atm.